CHG Pages Refactoring: Visuals
